% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{selectFac.MultiEFM}
\alias{selectFac.MultiEFM}
\title{Select the number of factors for MultiEFM}
\usage{
selectFac.MultiEFM(
  XList,
  q_max = 20,
  qs_max = 20,
  qd = NULL,
  n_threads = 4,
  sample_pairs = 0,
  verbose = TRUE,
  seed = 1
)
}
\arguments{
\item{XList}{A length-S list, where each component represents a data matrix for a specific study.}

\item{q_max}{an integer, specify the maximum number of study-shared factors to consider; default as 20.}

\item{qs_max}{an integer, specify the maximum number of study-specified factors to consider; default as 20.}

\item{qd}{an optional integer, specify the true number of combined factors if known in advance; default is NULL.}

\item{n_threads}{an integer, specify the number of threads for parallel computing; default as 4.}

\item{sample_pairs}{an integer, specify the number of sample pairs used for estimation; default as 0.}

\item{verbose}{a logical value, whether to output the estimation information.}

\item{seed}{an integer, specify the random seed for reproducibility; default as 1.}
}
\value{
return a list including the following components:
\item{hq}{the estimated number of study-shared factors;}
\item{hqs_vec}{an integer vector representing the estimated number of study-specific factors for each study;}
\item{eigvals_A}{the eigenvalues used for determining the shared factor number;}
\item{eig_ratio_A}{the eigenvalue ratios used for determining the shared factor number;}
\item{eigvals_B}{a list of eigenvalues used for determining the specific factor numbers;}
\item{eig_ratio_B}{a list of eigenvalue ratios used for determining the specific factor numbers.}
}
\description{
Automatically estimate the number of study-shared factors and study-specific factors based on the eigenvalue ratio method.
}
